test(SWArch - 015): restructurise page object model tree
