+++ pageObjects/shoppingCartPage/shoppingCartPage.js
